remove Django dependency in setup.py
This is to temporarily work around the issue with pip install not really evaluating version restrictions for dependencies correctly.
Because aldryn-addons is usually first listed in requirements.txt, pip will happily install the newest version of Django, even if there is a other package somewhere in requirements.txt, which as its own dependency specifies a different Django version.